Review: Move from _scalers.py to _modules.py
